Arreglé los permisos para agregar depas y categorias

+++ src/controllers/departamento.js
@@ -17,32 +17,32 @@ const obtenerDepartamentos = async (req, res) => {
     }
 }
 
-const obtenerDepartamento = async(req, res) => {
-    const { id } = req.empresa
-    const { nombre } = req.params
+// const obtenerDepartamento = async(req, res) => {
+//     const { id } = req.empresa
+//     const { nombre } = req.params
 
     
-    try {
-        const departamento = await Departamento.findOne({ empresa: id, nombre})
+//     try {
+//         const departamento = await Departamento.find({ empresa: id, nombre})
         
-        if(!departamento) {
-            return res.status(404).json({ msg: "No se encontró el departamento" })
+//         if(!departamento) {
+//             return res.status(404).json({ msg: "No se encontró el departamento" })
 
-        }
+//         }
 
-        if (departamento.categorias.length > 0) {
-            const departamentoCategorias = await Departamento.findOne({ empresa: id, nombre}).populate("categorias")
-            return res.status(200).json({ departamento: departamentoCategorias })
-        }
+//         if (departamento.categorias.length > 0) {
+//             const departamentoCategorias = await Departamento.findOne({ empresa: id, nombre}).populate("categorias")
+//             return res.status(200).json({ departamento: departamentoCategorias })
+//         }
 
-        return res.status(200).json({ departamento })
+//         return res.status(200).json({ departamento })
         
-    } catch (error) {
-        console.log(error)
-        return res.status(500).json({ error })
-    }
+//     } catch (error) {
+//         console.log(error)
+//         return res.status(500).json({ error })
+//     }
     
-}
+// }
 
 const crearDepartamento = async(req, res) => {
     const { nombre: nombreDepartamento, icon } = req.body
@@ -67,9 +67,24 @@ const crearDepartamento = async(req, res) => {
 const editarDepartamento = async(req, res) => {
     const { id } = req.params
     const { nombre, icon } = req.body
+    const { id: idEmpresa } = req.empresa
     try {
-        const departamento = await Departamento.findByIdAndUpdate(id, { nombre, icon }, { new: true})
+        const departamento = await Departamento.findById(id)
+            .populate({
+                path: 'empresa',
+                select: '_id'
+            })
+
+        if (!departamento) return res.status(404).json({ msg: "No se ha encontrado a ese departamento" });
+
+        if(idEmpresa !== departamento.empresa.id) return res.status(401).json({ msg: "No tienes permisos para editar este departamento" })
+
+        departamento.nombre = nombre
+        departamento.icon = icon
+        await departamento.save()
+
         return res.status(200).json({ departamento })
+
     }catch(error) {
         console.log(error)
         return res.status(500).json({ error })
@@ -78,14 +93,24 @@ const editarDepartamento = async(req, res) => {
 
 const eliminarDepartamento = async(req, res) => {
     const { id } = req.params
-    try {
-        
-        await Departamento.findByIdAndDelete(id)
+    const { id: idEmpresa } = req.empresa
+    try {        
+        const departamento = await Departamento.findById(id)
+            .populate({
+                path: 'empresa',
+                select: '_id'
+            })
+
+        if (!departamento) return res.status(404).json({ msg: "No se ha encontrado a ese departamento" });
+        if(idEmpresa !== departamento.empresa.id) return res.status(401).json({ msg: "No tienes permisos para eliminar este departamento" })
+
+        const eliminarDepto = await Departamento.findByIdAndDelete(id)
+
         return res.status(200).json({ msg: "Departamento eliminado correctamente" })
     }catch(error) {
-
+        console.log(error)
         return res.status(500).json({ error })
     }
 }
 
-export default { obtenerDepartamento, crearDepartamento, editarDepartamento, eliminarDepartamento , obtenerDepartamentos }
+export default { obtenerDepartamentos, crearDepartamento, editarDepartamento, eliminarDepartamento }
